requirements.txt
spotipy==2.23.0
youtube-search-python
discord.py